added test case
